+++ Sources/mtpx/Mtpx.swift
@@ -13,8 +13,8 @@ struct Mtpx: AsyncParsableCommand {
 			  mtpx --generate-completion-script zsh \
 			    > ~/.zsh/completions/_mtpx
 			""",
-		version: "0.4.0",
-		subcommands: [Transfer.self, Ls.self, DeviceCommand.self],
+		version: "0.5.0",
+		subcommands: [Transfer.self, Sync.self, Ls.self, DeviceCommand.self],
 		defaultSubcommand: Transfer.self
 	)
 }